Random cleanup
git-svn-id: svn://192.168.1.10/public/CoffeeMud@18303 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Common/GatheringSkill.java
@@ -88,7 +88,7 @@ public class GatheringSkill extends CommonSkill
 				final Quad<Room,Integer,short[],Long> oldRecord = roomSpamCounter.get(R);
 				if((oldRecord!=null)
 				&&((System.currentTimeMillis() > oldRecord.fourth.longValue())
-				  ||(oldRecord.second != R.myResource())))
+				  ||(oldRecord.second.intValue() != R.myResource())))
 				{
 					roomSpamCounter.remove(R);
 					return true;
@@ -118,7 +118,7 @@ public class GatheringSkill extends CommonSkill
 				final Quad<Room,Integer,short[],Long> oldRecord = roomSpamCounter.get(R);
 				if((oldRecord!=null)
 				&&((now > oldRecord.fourth.longValue())
-				  ||(oldRecord.second != R.myResource())))
+				  ||(oldRecord.second.intValue() != R.myResource())))
 					roomSpamCounter.remove(R);
 			}
 			for(final Iterator<Room> i=roomSpamCounter.keySet().iterator();i.hasNext();)
